navigation and loading of page body complete
